feat(eslint-plugin): add no-http-url rule
This commit introduces a new ESLint rule named "no-http-url". This rule
disallows the use of HTTP URLs in the code, suggesting the use of HTTPS
instead for better security. The rule is fixable and automatically
replaces "http://" with "https://". Tests for the rule have also been
added. The rule does not apply to localhost and 127.0.0.1 URLs.